refactor: Make extensions use MixDrop shared lib
