+++ src/en/animedao/build.gradle
@@ -11,11 +11,11 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-mixdrop-extractor'))
     implementation(project(':lib-mp4upload-extractor'))
     implementation(project(':lib-streamtape-extractor'))
     implementation(project(':lib-streamsb-extractor'))
     implementation(project(':lib-dood-extractor'))
-    implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
 }
