+++ src/en/vidembed/build.gradle
@@ -11,9 +11,9 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-mixdrop-extractor'))
     implementation(project(':lib-streamsb-extractor'))
     implementation(project(':lib-dood-extractor'))
-    implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
 }
