+++ src/es/doramasflix/build.gradle
@@ -10,12 +10,12 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-mixdrop-extractor'))
     implementation(project(':lib-streamtape-extractor'))
     implementation(project(':lib-okru-extractor'))
     implementation(project(':lib-streamsb-extractor'))
     implementation(project(':lib-dood-extractor'))
-    implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
-    implementation project(path: ':lib-voe-extractor')
+    implementation(project(':lib-voe-extractor'))
 }
 
 apply from: "$rootDir/common.gradle"
