+++ src/pt/megaflix/build.gradle
@@ -12,6 +12,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-mixdrop-extractor'))
     implementation(project(":lib-streamsb-extractor"))
     implementation(project(":lib-streamtape-extractor"))
     // for mixdrop and megaflix
